+++ CHANGELOG.md
@@ -10,7 +10,7 @@ phpSysInfo 3.1.13
  - [UPD] Romanian Translation ro.xml
 
  - [ADD] SHOW_CPUINFO_EXPANDED option
- - [ADD] Machine detection on Linux, Android and Darwin systems
+ - [ADD] Machine detection on WINNT, Linux, Android and Darwin systems
 
 
 phpSysInfo 3.1.12